+++ datajoint/base_relation.py
@@ -7,13 +7,13 @@ import pymysql
 import logging
 import warnings
 from pymysql import OperationalError, InternalError, IntegrityError
-from . import config, DataJointError
+from . import config
 from .declare import declare
 from .relational_operand import RelationalOperand
 from .blob import pack
 from .utils import user_choice
 from .heading import Heading
-from .settings import server_error_codes
+from .errors import server_error_codes, DataJointError, DuplicateError
 from . import __version__ as version
 
 logger = logging.getLogger(__name__)
@@ -44,7 +44,8 @@ class BaseRelation(RelationalOperand):
         if not self._heading:  # lazy loading of heading
             if self.connection is None:
                 raise DataJointError(
-                    'DataJoint class is missing a database connection. Missing class decorator?')
+                    'DataJoint class is missing a database connection. '
+                    'Missing schema decorator on the class? (e.g. @schema)')
             else:
                 self._heading.init_from_database(self.connection, self.database, self.table_name)
         return self._heading
@@ -287,10 +288,12 @@ class BaseRelation(RelationalOperand):
                 elif err.args[0] == server_error_codes['unknown column']:
                     # args[1] -> Unknown column 'extra' in 'field list'
                     raise DataJointError(
-                        '{} : To ignore extra fields, set ignore_extra_fields=True in insert.'.format(err.args[1])) from None
+                        '{} : To ignore extra fields, set ignore_extra_fields=True in insert.'.format(err.args[1])
+                    ) from None
                 elif err.args[0] == server_error_codes['duplicate entry']:
-                    raise DataJointError(
-                        '{} : To ignore duplicate entries, set skip_duplicates=True in insert.'.format(err.args[1])) from None
+                    raise DuplicateError(
+                        '{} : To ignore duplicate entries, set skip_duplicates=True in insert.'.format(err.args[1])
+                    ) from None
                 else:
                     raise
 
@@ -445,7 +448,9 @@ class BaseRelation(RelationalOperand):
             This does not yet work for aliased foreign keys.
         """
         if context is None:
-            context = inspect.currentframe().f_back.f_globals
+            frame = inspect.currentframe().f_back
+            context = dict(frame.f_globals, **frame.f_locals)
+            del frame
         if self.full_table_name not in self.connection.dependencies:
             self.connection.dependencies.load()
         parents = self.parents()